+++ lib/type_scopes/numeric.rb
@@ -9,8 +9,8 @@ class TypeScopes::Numeric < TypeScopes
     append_scope(model, :"#{name}_from", lambda { |value| where(column.gteq(value)) })
     append_scope(model, :"#{name}_above", lambda { |value| where(column.gt(value)) })
     append_scope(model, :"#{name}_below", lambda { |value| where(column.lt(value)) })
-    append_scope(model, :"#{name}_between", lambda { |from, to| where(name => from..to) })
-    append_scope(model, :"#{name}_not_between", lambda { |from, to| where.not(name => from..to) })
+    append_scope(model, :"#{name}_between", lambda { |*args| where(name => TypeScopes.args_to_range(*args)) })
+    append_scope(model, :"#{name}_not_between", lambda { |*args| where.not(name => TypeScopes.args_to_range(*args)) })
     append_scope(model, :"#{name}_within", lambda { |from, to| where(column.gt(from)).where(column.lt(to)) })
     append_scope(model, :"#{name}_not_within", lambda { |from, to| where(column.lteq(from).or(column.gteq(to))) })
   end