+++ lib/type_scopes/time.rb
@@ -11,8 +11,8 @@ class TypeScopes::Time < TypeScopes
     append_scope(model, :"#{short_name}_from", lambda { |date| where(column.gteq(date)) })
     append_scope(model, :"#{short_name}_after", lambda { |date| where(column.gt(date)) })
     append_scope(model, :"#{short_name}_before", lambda { |date| where(column.lt(date)) })
-    append_scope(model, :"#{short_name}_between", lambda { |from, to| where(name => from..to) })
-    append_scope(model, :"#{short_name}_not_between", lambda { |from, to| where.not(name => from..to) })
+    append_scope(model, :"#{short_name}_between", lambda { |*args| where(name => TypeScopes.args_to_range(*args)) })
+    append_scope(model, :"#{short_name}_not_between", lambda { |*args| where.not(name => TypeScopes.args_to_range(*args)) })
     append_scope(model, :"#{short_name}_within", lambda { |from, to| where(column.gt(from)).where(column.lt(to)) })
     append_scope(model, :"#{short_name}_not_within", lambda { |from, to| where(column.lteq(from).or(column.gteq(to))) })
   end